requirements.txt
rivescript==1.15.0
cairosvg==2.5.1
python-chess==0.31.4